Move Formatters to Common views folder
