+++ miataru/miataru/views/Common/Formatters.swift
@@ -1,0 +1,47 @@
+/*
+ * Copyright (c) 2013-2025, Daniel Kirstenpfad, www.miataru.com
+ *
+ * Formatters.swift
+ * miataru
+ *
+ * Created by Daniel Kirstenpfad on 27.09.24.
+ */
+
+import Foundation
+
+enum Formatters {
+    static let relativeDateTime: RelativeDateTimeFormatter = {
+        let formatter = RelativeDateTimeFormatter()
+        return formatter
+    }()
+
+    static let time: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateStyle = .none
+        formatter.timeStyle = .medium
+        return formatter
+    }()
+
+    static let duration: DateComponentsFormatter = {
+        let formatter = DateComponentsFormatter()
+        formatter.unitsStyle = .short
+        return formatter
+    }()
+
+    static let number: NumberFormatter = {
+        let formatter = NumberFormatter()
+        formatter.maximumFractionDigits = 0
+        formatter.minimumFractionDigits = 0
+        return formatter
+    }()
+
+    static let measurement: MeasurementFormatter = {
+        let formatter = MeasurementFormatter()
+        formatter.locale = .current
+        formatter.unitStyle = .short
+        formatter.unitOptions = .providedUnit
+        formatter.numberFormatter = Formatters.number
+        return formatter
+    }()
+}
+
