+++ miataru/miataru/views/iPhone/Devices Views/iPhone_DeviceNavigationView.swift
@@ -478,8 +478,7 @@ struct iPhone_DeviceNavigationView: View {
                 let response = try await MKDirections(request: request).calculate()
                 if let first = response.routes.first {
                     route = first
-                    let formatter = DateComponentsFormatter()
-                    formatter.unitsStyle = .short
+                    let formatter = Formatters.duration
                     travelTime = formatter.string(from: first.expectedTravelTime)
                     distanceText = formattedDistance(first.distance)
                     // Remember inputs used for this route calculation
@@ -756,8 +755,7 @@ extension iPhone_DeviceNavigationView {
             ) {
                 // Apply cached route and keep display fields in sync
                 route = cached.route
-                let formatter = DateComponentsFormatter()
-                formatter.unitsStyle = .short
+                let formatter = Formatters.duration
                 travelTime = formatter.string(from: cached.route.expectedTravelTime)
                 distanceText = formattedDistance(cached.route.distance)
                 // Align last-route inputs so existing change detection logic works
@@ -788,14 +786,7 @@ extension iPhone_DeviceNavigationView {
             unit = .miles
         }
         let measurement = Measurement(value: value, unit: unit)
-        let formatter = MeasurementFormatter()
-        formatter.locale = .current
-        formatter.unitStyle = .short
-        formatter.unitOptions = .providedUnit
-        let numberFormatter = NumberFormatter()
-        numberFormatter.maximumFractionDigits = 0
-        numberFormatter.minimumFractionDigits = 0
-        formatter.numberFormatter = numberFormatter
+        let formatter = Formatters.measurement
         return formatter.string(from: measurement)
     }
